Split upload task so changed files can be traced with dry run
