+++ node-src/ui/tasks/uploadE2E.stories.ts
@@ -1,18 +1,13 @@
 import task from '../components/task';
 import {
-  bailed,
   dryRun,
   failed,
   finalizing,
-  hashing,
   initial,
   invalid,
   starting,
   success,
-  traced,
-  tracing,
   uploading,
-  validating,
 } from './upload';
 
 export default {
@@ -26,35 +21,12 @@ export const Initial = () => initial(ctx);
 
 export const DryRun = () => dryRun(ctx);
 
-export const Validating = () => validating(ctx);
-
 export const Invalid = () =>
   invalid({
     ...ctx,
     sourceDir: '/var/folders/h3/ff9kk23958l99z2qbzfjdlxc0000gn/T/chromatic-20036LMP9FAlLEjpu',
     buildLogFile: '/var/folders/h3/ff9kk23958l99z2qbzfjdlxc0000gn/T/build-storybook.log',
   } as any);
-
-export const Tracing = () =>
-  tracing({ ...ctx, git: { changedFiles: Array.from({ length: 3 }) } } as any);
-
-export const BailedPackageFile = () =>
-  bailed({ ...ctx, turboSnap: { bailReason: { changedPackageFiles: ['package.json'] } } } as any);
-
-export const BailedLockfile = () =>
-  bailed({ ...ctx, turboSnap: { bailReason: { changedPackageFiles: ['yarn.lock'] } } } as any);
-
-export const BailedSiblings = () =>
-  bailed({
-    ...ctx,
-    turboSnap: {
-      bailReason: { changedStorybookFiles: ['.storybook/preview.js', '.storybook/otherfile.js'] },
-    },
-  } as any);
-
-export const Traced = () => traced({ ...ctx, onlyStoryFiles: Array.from({ length: 5 }) } as any);
-
-export const Hashing = () => hashing(ctx);
 
 export const Starting = () => starting(ctx);
